chore: add what we can fetch asset-wise to the build.rs
